requirements.txt
beautifulsoup4==4.10.0
Pillow==9.0.1
requests==2.25.1
types-beautifulsoup4==4.10
types-Pillow==8.3
types-requests==2.25
